requirements.txt
torch>=2.5.1+cu118
torchvision>=0.20.1+cu118
numpy
matplotlib
Pillow
scikit-learn
tqdm
pyqt5
opencv